Handle platform-specific DuckDB enum types
Bindgen emits DuckDB C enums as c_int on Windows and c_uint on Unix.

Cast at FFI boundaries while preserving u32-facing logical type IDs.

+++ crates/duckdb/src/statement.rs
@@ -2038,7 +2038,10 @@ mod test {
 
         let logical_type = stmt.column_logical_type(0);
         assert_eq!(logical_type.id(), LogicalTypeId::Variant);
-        assert_eq!(logical_type.raw_id(), crate::ffi::DUCKDB_TYPE_DUCKDB_TYPE_VARIANT);
+        assert_eq!(
+            logical_type.raw_id(),
+            crate::ffi::DUCKDB_TYPE_DUCKDB_TYPE_VARIANT as u32
+        );
         assert_eq!(logical_type.num_children(), 0);
         assert_eq!(format!("{logical_type:?}"), "Variant");
 
@@ -2052,7 +2055,10 @@ mod test {
 
         let logical_type = stmt.column_logical_type(0);
         assert_eq!(logical_type.id(), LogicalTypeId::Geometry);
-        assert_eq!(logical_type.raw_id(), crate::ffi::DUCKDB_TYPE_DUCKDB_TYPE_GEOMETRY);
+        assert_eq!(
+            logical_type.raw_id(),
+            crate::ffi::DUCKDB_TYPE_DUCKDB_TYPE_GEOMETRY as u32
+        );
         assert_eq!(logical_type.geometry_crs(), None);
         assert_eq!(format!("{logical_type:?}"), "Geometry");
